+++ docs/changelog.md
@@ -1,5 +1,34 @@
 # Change Log
 
+## v1.0.3
+
+### Enhancements
+
+* [#533](https://github.com/netboxlabs/netbox-branching/issues/533) - Avoid instantiating objects when verifying their existence
+
+### Bug Fixes
+
+* [#542](https://github.com/netboxlabs/netbox-branching/issues/542) - Branch migrations should not generate object change records
+
+---
+
+## v1.0.2
+
+### Bug Fixes
+
+* [#539](https://github.com/netboxlabs/netbox-branching/issues/539) - Fix branch upgrade migration failures from NetBox 4.4 to 4.6
+
+---
+
+## v1.0.1
+
+### Bug Fixes
+
+* [#529](https://github.com/netboxlabs/netbox-branching/issues/529) - Allow specifying migrations to skip via `fake_on_branch` to handle `RunPython` and `AlterModelOptions` operations
+
+---
+
+
 ## v1.0.0
 
 ### Enhancements